requirements.txt
# Base requirements - works for development and cloud training
torch>=2.0.0
torchvision>=0.15.0

# RL Framework
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Logging
wandb>=0.15.0
tensorboard>=2.14.0
matplotlib>=3.7.0

# Vision
opencv-python>=4.8.0
ultralytics>=8.0.0

# Utilities
numpy>=1.24.0
PyYAML>=6.0

# SO-100 SDK (uncomment when arm is connected)
# git+https://github.com/huggingface/lerobot.git

# Simulator (uncomment based on choice)
# isaacgym  # Requires manual install from NVIDIA
# genesis-world>=0.2.0  # Alternative: pip install genesis-world

# Development
pytest>=7.4.0
